Potential Get-AzureRmVpnClientPackage issue
After performing this script many times for Azure Government Web Apps, we've come to find that the Get-AzureRmVpnClientPackage returns a string with actual double quotes on its value, thus making the URL incorrect. When the URL incorrectly passed to the Set-AzureRmResource for the gateway, it stores is blank without any error messages. So we had to add a check that basically removes the leading and trailing double-quotes from the string returned from Get-AzureRmVpnClientPackage. This edit is proposed in the "Resource Manager VNet App Service integration script" PowerShell script. I admit it can be written better since I'm only checking for the leading double-quote and assuming the trailing one.